requirements.txt
joblib
bentoml
torchvision
tqdm
pip-chill
wincertstore


-e .
